+++ src/classes/Settings.ts
@@ -1,3 +1,4 @@
+const storage = require("uxp").storage;
 const localFileSystem = require("uxp").storage.localFileSystem;
 
 export class Settings{
@@ -14,13 +15,31 @@ export class Settings{
 		return folder;
 	}
 
-	public static async saveSettings(object: any): Promise<void> {
-		if (!this.loaded) {
+	public static async saveSettings(object: any): Promise<void> {		
+		const folder = await this.settingsFolder();
+		await this._saveSettings(object, folder, this.settingsFilename);
+	}
+
+	public static async saveSettingsWithDialog(object: any): Promise<void>{
+		const file = await localFileSystem.getFileForSaving(Date.now() + ".json", {
+			types: ["json"],
+			//initialLocation: await this.settingsFolder()
+		});
+		if (!file) {
 			return
 		}
 		const data = JSON.stringify(object, null, "\t");
-		
-		const folder = await this.settingsFolder();
+		await file.write(data, {
+			append: false
+		});
+	}
+
+	private static async _saveSettings(object: any, folder: any, name:string): Promise<void>{
+		if (!this.loaded) {
+			return
+		}
+
+		const data = JSON.stringify(object, null, "\t");
 		console.log(folder);
 		const created = await folder.createFile(this.settingsFilename, {
 			overwrite: true
@@ -30,7 +49,7 @@ export class Settings{
 		});
 	}
 
-	public static async loadSettings(): Promise<any|null>{
+	public static async importState(): Promise<any|null>{
 		const folder = await this.settingsFolder();
 		const entries:any[] = await folder.getEntries();
 		const found = entries.find(file => file.name === this.settingsFilename);
@@ -40,6 +59,25 @@ export class Settings{
 		const data: string = await found.read();
 		try {
 			const result = JSON.parse(data);			
+			return result;
+		} catch (e) {
+			console.log("Error - with reading of settings!");
+			return null;
+		}
+	}
+
+	public static async importStateWithDialog(): Promise<any|null>{
+		const files = await localFileSystem.getFileForOpening({
+			types: ["json"],
+			allowMultiple: true
+			//initialLocation: await this.settingsFolder()
+		});
+		if (!files || !files.length) {
+			return null
+		}
+		const data: string = await files[0].read();
+		try {
+			const result = JSON.parse(data);
 			return result;
 		} catch (e) {
 			console.log("Error - with reading of settings!");